+++ src/common/IConnection.ts
@@ -2,8 +2,9 @@ export interface IConnection {
   readonly host: string;
   readonly user: string;
   password?: string;
-  readonly port: string;
+  readonly port: number;
   readonly database?: string;
   multipleStatements?: boolean;
   readonly certPath?: string;
+  ssl?: any;
 }